requirements.txt
Flask==1.1.1
line-bot-sdk==1.16.0
schedule==0.6.0
requests==2.23.0
gspread==3.4.2
oauth2client==4.1.3